Updated the version, so it uses the newest database methods.
