+++ src/main/java/com/xxxx/example/servletPostPostGateway.java
@@ -1,6 +1,7 @@
 package com.xxxx.example;
 
 import database_connection.Database;
+import database_connection.DatabaseInsert;
 import post_post_use_case.ReturnAsPost;
 import post_reply_user.Post;
 
@@ -17,10 +18,10 @@ public class servletPostPostGateway extends HttpServlet {
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String text = (String) req.getParameter("text");
         String userName = (String) req.getParameter("userName");
-        Database myDatabase = new Database("", "DatingAppStaging");
+        DatabaseInsert myDatabase = new DatabaseInsert("", "DatingAppStaging");
         ReturnAsPost returnPost = new ReturnAsPost();
         Post post = returnPost.returnPost(userName, text);
-        myDatabase.insert_post(post);
+        myDatabase.insertPost(post);
         myDatabase.close();
         req.getRequestDispatcher("/postPostResponse").forward(req,resp);
     }
