+++ src/main/java/com/xxxx/example/servletRegisterGateway.java
@@ -1,6 +1,8 @@
 package com.xxxx.example;
 
 import database_connection.Database;
+import database_connection.DatabaseInsert;
+import database_connection.DatabaseRead;
 import user_exist_use_case.UserExistCheck;
 import post_reply_user.CommonUser;
 import register_use_case.ReturnAsUser;
@@ -17,17 +19,19 @@ import java.io.IOException;
 public class servletRegisterGateway extends HttpServlet {
     private boolean check;
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        Database myDatabase = new Database("", "DatingAppStaging");
+        DatabaseRead myDatabase = new DatabaseRead("", "DatingAppStaging");
         String username = req.getParameter("username");
-        CommonUser returnedUsername = myDatabase.find_user_by_id(username);
+        CommonUser returnedUsername = myDatabase.findUserById(username);
         String password = req.getParameter("password");
         UserExistCheck userCheck = new UserExistCheck();
         boolean check = userCheck.userExistCheck(returnedUsername);
+        myDatabase.close();
         if (!check) {
             ReturnAsUser returnUser = new ReturnAsUser();
             CommonUser user = returnUser.returnUser(username, password, null);
-            myDatabase.insert_user(user);
-            myDatabase.close();
+            DatabaseInsert databaseInsert = new DatabaseInsert("", "DatingAppStaging");
+            databaseInsert.insertUser(user);
+            databaseInsert.close();
             req.getRequestDispatcher("/registerResponse").forward(req, resp);
         } else {
             req.getRequestDispatcher("/registerResponseUserExist").forward(req, resp);
