+++ src/main/java/com/xxxx/example/servletSendReplyGateway.java
@@ -1,6 +1,7 @@
 package com.xxxx.example;
 
 import database_connection.Database;
+import database_connection.DatabaseInsert;
 import post_reply_user.Reply;
 import send_reply_use_case.ReturnAsReply;
 
@@ -19,10 +20,10 @@ public class servletSendReplyGateway extends HttpServlet {
         String text = req.getParameter("text");
         String username = req.getParameter("username");
         String id = req.getParameter("id");
-        Database myDatabase = new Database("", "DatingAppStaging");
+        DatabaseInsert myDatabase = new DatabaseInsert("", "DatingAppStaging");
         ReturnAsReply re = new ReturnAsReply();
         Reply reply = re.returnReply(username, id, text);
-        myDatabase.insert_reply(reply);
+        myDatabase.insertReply(reply);
         myDatabase.close();
         req.getRequestDispatcher("/sendReplyResponse").forward(req,resp);
     }
