--- src/main/java/popup_window/index.html
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Login</title>
-    <link rel="stylesheet" href="popupwindow.css" />
-  </head>
-  <body>
-    <button id="open_btn" class="btn">Login</button>
-    <!-- start of pop up window -->
-    <div id="background" class="back">
-      <div id="div1" class="content">
-        <div id="close">
-          <span id="close-button">×</span>
-          <h2>Login / Sign Up</h2>
-        </div>
-        <div id="div2">
-          <h3>Password Login</h3>
-          <form class="form">
-            <label class="account__label"
-              >account number
-              <input
-                class="account"
-                type="text"
-                placeholder="Please input your account number"
-              />
-            </label>
-            <label class="password__label"
-              >password
-              <input
-                class="password"
-                type="password"
-                placeholder="Please input your password"
-              />
-            </label>
-          </form>
-          <div class="buttons">
-            <button class="button">LOGIN</button>
-            <button class="button button--register">REGISTER</button>
-          </div>
-        </div>
-        <div class="footer">
-          <span>Others</span>
-          <img
-            class="social__icon"
-            src="../../../../../../Desktop/myproject/chatting-app-servlet/course-project-dating-app-team/images/Icon-facebook.svg"
-            alt="facebook"
-          />
-          <img
-            class="social__icon"
-            src="../../../../../../Desktop/myproject/chatting-app-servlet/course-project-dating-app-team/images/Icon-instagram.svg"
-            alt="Instagram"
-          />
-          <img
-            class="social__icon"
-            src="../../../../../../Desktop/myproject/chatting-app-servlet/course-project-dating-app-team/images/Icon-twitter.svg"
-            alt="twitter"
-          />
-        </div>
-      </div>
-    </div>
-    <script src="popupwindow.js"></script>
-  </body>
-</html>
