--- src/main/java/popup_window/popupwindow.css
@@ -1,122 +1,0 @@
-body {
-  font-family: Arial, Helvetica, sans-serif;
-  font-size: 17px;
-  text-align: 1.5;
-}
-
-#open_btn {
-  background: beige;
-}
-
-#background {
-  display: none;
-  position: fixed;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-}
-
-#div1 {
-  background: #eeeeee;
-  width: 70%;
-  z-index: 1;
-  margin: 12% auto;
-  overflow: auto;
-}
-
-span {
-  color: white;
-  padding-top: 12px;
-  cursor: pointer;
-  padding-right: 15px;
-}
-
-#div2 {
-  background: #eeeeee;
-  height: 300px;
-  padding: 0 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-#close {
-  padding: 5px;
-  background: beige;
-}
-
-#close-button {
-  float: right;
-  font-size: 30px;
-}
-
-#foot {
-  padding: 5px 0;
-  text-align: center;
-  background: beige;
-  color: white;
-}
-
-h2 {
-  margin: 10px 0;
-  color: pink;
-  padding-left: 15px;
-  text-align: center;
-}
-
-h3 {
-  margin: 0 0px;
-  padding-top: 15px;
-  /* text-align: center; */
-}
-.form {
-  margin-top: 4rem;
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-  justify-content: center;
-  align-items: flex-start;
-  /* position: relative;
-  left: 35%; */
-}
-
-.account__label,
-.password__label {
-  width: 100%;
-}
-.buttons {
-  display: flex;
-  /* width: 50%; */
-  gap: 1rem;
-  /* position: relative;
-  left: 35%; */
-  margin-top: 2rem;
-}
-.button {
-  background-color: pink;
-  color: white;
-  width: 8rem;
-  height: 40px;
-  border-radius: 4px;
-  border: 1px solid white;
-}
-.button:hover {
-  background-color: salmon;
-}
-.button--register {
-  background-color: white;
-  color: black;
-}
-.footer {
-  display: flex;
-  background-color: pink;
-  justify-content: center;
-  align-items: baseline;
-  padding: 0.5rem 0;
-  gap: 3rem;
-}
-.social__icon {
-  height: 1rem;
-}
