+++ src/main/java/post_reply_user/CommonUser.java
@@ -72,6 +72,3 @@ public class CommonUser extends User{
 //    //the user delete a post
 
 }
-
-
-
