+++ src/main/java/post_reply_user/User.java
@@ -30,6 +30,4 @@ public abstract class User {
     }
 
 
-
-
 }
